pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smiles-bd"
version = "0.1.0"
description = "Minimal SMILES Block Diffusion (MDLM/SUBS) with prefix-conditional sampling"
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["torch>=2.1", "numpy", "pyyaml>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["smiles_bd*"]

[tool.setuptools.package-data]
"smiles_bd" = []

[tool.setuptools]
package-dir = {"" = "src"}
